Move tests between test_runtime_env_working_dir_2 and test_runtime_env_working_dir_3
