+++ python/ray/tests/test_runtime_env_working_dir_2.py
@@ -3,16 +3,16 @@ from pathlib import Path
 import sys
 import time
 import tempfile
-from unittest import mock
 
 import pytest
 from pytest_lazyfixture import lazy_fixture
 from ray._private.test_utils import run_string_as_driver
 
 import ray
-from ray._private.test_utils import chdir, check_local_files_gced
+from ray._private.test_utils import wait_for_condition, chdir, check_local_files_gced
 from ray._private.runtime_env import RAY_WORKER_DEV_EXCLUDES
 from ray._private.runtime_env.packaging import GCS_STORAGE_MAX_SIZE
+from ray.exceptions import GetTimeoutError
 
 # This test requires you have AWS credentials set up (any AWS credentials will
 # do, this test only accesses a public bucket).
@@ -21,19 +21,6 @@ from ray._private.runtime_env.packaging import GCS_STORAGE_MAX_SIZE
 # Calling `test_module.one()` should return `2`.
 # If you find that confusing, take it up with @jiaodong...
 S3_PACKAGE_URI = "s3://runtime-env-test/test_runtime_env.zip"
-
-
-@pytest.fixture(scope="class")
-def URI_cache_10_MB():
-    with mock.patch.dict(
-        os.environ,
-        {
-            "RAY_RUNTIME_ENV_WORKING_DIR_CACHE_SIZE_GB": "0.01",
-            "RAY_RUNTIME_ENV_PY_MODULES_CACHE_SIZE_GB": "0.01",
-        },
-    ):
-        print("URI cache size set to 0.01 GB.")
-        yield
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="Fail to create temp dir.")
@@ -298,39 +285,134 @@ def test_default_large_cache(start_cluster, option: str, source: str):
     assert not check_local_files_gced(cluster)
 
 
-# Set scope to "class" to force this to run before start_cluster, whose scope
-# is "function".  We need these env vars to be set before Ray is started.
-@pytest.fixture(scope="class")
-def skip_local_gc():
-    with mock.patch.dict(
-        os.environ,
+@pytest.mark.skipif(
+    os.environ.get("CI") and sys.platform != "linux",
+    reason="Requires PR wheels built in CI, so only run on linux CI machines.",
+)
+@pytest.mark.parametrize(
+    "ray_start_cluster",
+    [
         {
-            "RAY_RUNTIME_ENV_SKIP_LOCAL_GC": "1",
+            "num_nodes": 1,
+            "_system_config": {
+                "num_workers_soft_limit": 0,
+            },
         },
+        {
+            "num_nodes": 1,
+            "_system_config": {
+                "num_workers_soft_limit": 5,
+            },
+        },
+        {
+            "num_nodes": 1,
+            "_system_config": {
+                "num_workers_soft_limit": 0,
+                # this delay will make worker start slow and time out
+                "testing_asio_delay_us": "InternalKVGcsService.grpc_server"
+                ".InternalKVGet=2000000:2000000",
+                "worker_register_timeout_seconds": 1,
+            },
+        },
+        {
+            "num_nodes": 1,
+            "_system_config": {
+                "num_workers_soft_limit": 5,
+                # this delay will make worker start slow and time out
+                "testing_asio_delay_us": "InternalKVGcsService.grpc_server"
+                ".InternalKVGet=2000000:2000000",
+                "worker_register_timeout_seconds": 1,
+            },
+        },
+    ],
+    indirect=True,
+)
+@pytest.mark.skipif(sys.platform == "win32", reason="Fail to create temp dir.")
+@pytest.mark.parametrize("option", ["working_dir", "py_modules"])
+def test_task_level_gc(runtime_env_disable_URI_cache, ray_start_cluster, option):
+    """Tests that task-level working_dir is GC'd when the worker exits."""
+
+    cluster = ray_start_cluster
+
+    soft_limit_zero = False
+    worker_register_timeout = False
+    system_config = cluster.list_all_nodes()[0]._ray_params._system_config
+    if (
+        "num_workers_soft_limit" in system_config
+        and system_config["num_workers_soft_limit"] == 0
     ):
-        print("RAY_RUNTIME_ENV_SKIP_LOCAL_GC enabled.")
-        yield
+        soft_limit_zero = True
+    if (
+        "worker_register_timeout_seconds" in system_config
+        and system_config["worker_register_timeout_seconds"] != 0
+    ):
+        worker_register_timeout = True
 
+    @ray.remote
+    def f():
+        import test_module
 
-class TestSkipLocalGC:
-    @pytest.mark.parametrize("source", [lazy_fixture("tmp_working_dir")])
-    def test_skip_local_gc_env_var(self, skip_local_gc, start_cluster, source):
-        cluster, address = start_cluster
-        ray.init(address, namespace="test", runtime_env={"working_dir": source})
+        test_module.one()
 
-        @ray.remote
-        class A:
-            def test_import(self):
-                import test_module
+    @ray.remote(num_cpus=1)
+    class A:
+        def check(self):
+            import test_module
 
-                test_module.one()
+            test_module.one()
 
-        a = A.remote()
-        ray.get(a.test_import.remote())  # Check working_dir was downloaded
+    if option == "working_dir":
+        runtime_env = {"working_dir": S3_PACKAGE_URI}
+    else:
+        runtime_env = {"py_modules": [S3_PACKAGE_URI]}
 
-        ray.shutdown()
+    # Note: We should set a bigger timeout if downloads the s3 package slowly.
+    get_timeout = 10
 
-        time.sleep(1)  # Give time for GC to potentially happen
+    # Start a task with runtime env
+    if worker_register_timeout:
+        with pytest.raises(GetTimeoutError):
+            ray.get(f.options(runtime_env=runtime_env).remote(), timeout=get_timeout)
+    else:
+        ray.get(f.options(runtime_env=runtime_env).remote())
+    if soft_limit_zero or worker_register_timeout:
+        # Wait for worker exited and local files gced
+        wait_for_condition(lambda: check_local_files_gced(cluster))
+    else:
+        # Local files should not be gced because of an enough soft limit.
+        assert not check_local_files_gced(cluster)
+
+    # Start a actor with runtime env
+    actor = A.options(runtime_env=runtime_env).remote()
+    if worker_register_timeout:
+        with pytest.raises(GetTimeoutError):
+            ray.get(actor.check.remote(), timeout=get_timeout)
+        # Wait for worker exited and local files gced
+        wait_for_condition(lambda: check_local_files_gced(cluster))
+    else:
+        ray.get(actor.check.remote())
+        assert not check_local_files_gced(cluster)
+
+    # Kill actor
+    ray.kill(actor)
+    if soft_limit_zero or worker_register_timeout:
+        # Wait for worker exited and local files gced
+        wait_for_condition(lambda: check_local_files_gced(cluster))
+    else:
+        # Local files should not be gced because of an enough soft limit.
+        assert not check_local_files_gced(cluster)
+
+    # Start a task with runtime env
+    if worker_register_timeout:
+        with pytest.raises(GetTimeoutError):
+            ray.get(f.options(runtime_env=runtime_env).remote(), timeout=get_timeout)
+    else:
+        ray.get(f.options(runtime_env=runtime_env).remote())
+    if soft_limit_zero or worker_register_timeout:
+        # Wait for worker exited and local files gced
+        wait_for_condition(lambda: check_local_files_gced(cluster))
+    else:
+        # Local files should not be gced because of an enough soft limit.
         assert not check_local_files_gced(cluster)
